Client+Server: New Touchpad component

+++ server/PcRemoteServer/Services/WindowsInputService.cs
@@ -6,11 +6,11 @@ namespace PcRemoteServer.Services
     public class WindowsInputService : IInputService
     {
         [DllImport("user32.dll")]
-        private static extern void mouse_event(uint dwFlags, uint dx, uint dy, uint dwData, int dwExtraInfo);
+        private static extern void mouse_event(uint dwFlags, int dx, int dy, uint dwData, int dwExtraInfo);
         [DllImport("user32.dll")]
         private static extern void keybd_event(byte bVk, byte bScan, uint dwFlags, int dwExtraInfo);
 
-        private const uint MOUSEEVENTF_MOVE = 0x0001;
+        private const int MOUSEEVENTF_MOVE = 0x0001;
         private const uint MOUSEEVENTF_LEFTDOWN = 0x0002;
         private const uint MOUSEEVENTF_LEFTUP = 0x0004;
         private const uint MOUSEEVENTF_RIGHTDOWN = 0x0008;
@@ -30,7 +30,7 @@ namespace PcRemoteServer.Services
 
         public void MoveMouse(MouseMoveCommand command)
         {
-            mouse_event(MOUSEEVENTF_MOVE, (uint)command.Dx, (uint)command.Dy, 0, 0);
+            mouse_event(MOUSEEVENTF_MOVE, command.Dx, command.Dy, 0, 0);
         }
 
         public void PressKey(KeyCode keyCode)
